+++ src/test/java/spaceinvaders/spaceinvaders_framework/BulletTest.java
@@ -7,7 +7,9 @@ import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
 
-import Bullet.Bullet;
+import bullet.Bullet;
+import spaceinvaders.spaceinvaders_framework.Game;
+import spaceinvaders.spaceinvaders_framework.SpriteSheet;
 
 /**
  * BulletTest tests all the methods in the bullet class.
